cicd: travis: Deploy wheels
Wheels are deployed from master:
- If there is a tag then deployed to both pypi-test and
pypi.  The expectation is that this is done on a commit where the module
version specified in VERSION is incremented.
- Whithout a tag, then deployed to pypi-tst only.  The expectation is
that this is happening during incremental, between-releases activity on
master.  As such, the pipeline will insert / auto-increment a fourth grouping in
the package version:  for example 0.0.1.95.