repo: nanodbc: catch-up submodule to upstream
Change submodule to track branch 'cyanodbc-0.0.2' in our forked nanodbc repo. So
that once we release v0.0.2 we can freeze the branch we are tracking in
the forked repo - and not update it again.  Post v0.0.2 release, we
would need to start tracking a branch 'cyanodbc-0.0.3' in the forked
repo, and so on.